Lookup instance methods on the constant first
Looking up methods on the helper class has a problem since Rails creates
proxy methods (with the same name) for some helper methods on the helper
module. This messes up the generated parameters and prevents us from
ever getting the correct signature of the actual method.

Instead, if we look for the instance method on the `constant` first,
then we can return the method that we found for processing, since
ultimately we know that the parameters and the parameter types will that
of the method on the constant.

If the method does not exist on the constant, we fallback to looking up
the instance method on the helper module.

This has a slight chance of generating the wrong output if a helper
method defined in a `helper {}` and a method defined on the controller
end up having the same name but are totally unrelated. To prevent that
we could check for the source location of the helper instance method and
if it matches the location in `abstract_controller/helpers.rb` where
`helper_method` proxies are created, only then access the method on the
controller.