Make sure we only access method on constant if it is proxied
Rails stores the list of method names that it has setup proxies for in
the `_helper_methods` attribute, so instead of always trying to match
methods on the constant by name, we can check if the method name is in
that list and if so access the method on the controller instead of the
proxy on the helper module.

This solves the problem pointed out in the previous commit about the
edge case where blindly looking up by name might lead to generating the
wrong output.