Map generic types to T.untyped in custom columns
We end up not having access to generic type variables for a given type
because of type erasure by Sorbet runtime. Moreover, we don't even know
how many type variables a generic type will accept. However, just
generating a bare type name for the generic type will result in
type-checking errors since a type without type parameters supplied is
not valid.

As a result, the only and best thing we can do in those cases is to map
those to `T.untyped`.