Improve eager loading
We haven't been eager loading Zeitwerk dependencies so far, which was
preventing us from discovering some modules that we could generate RBIs
for. For example, `ActiveTax` uses Zeitwerk to do its autoloading and
has a `FrozenRecord` subclass `ActiveTax::TaxAuthority`. We were not
generating an RBI for that class since it was not eager loaded while we
were eager loading the Rails app.

It turns out, Rails app finishes its eager loading in an initializer
which does a little bit more than just eager loading all the Rails
autoloaders. That includes eager loading all of the Zeitwerk
dependencies as well. Thus, we need to replicate this logic in our eager
loading of the Rails application as well.

Ref: https://github.com/rails/rails/blob/master/railties/lib/rails/application/finisher.rb#L127-L136