Add test to verify Zeitwerk eager loading
Made the `baz` gem autoloaded via Zeitwerk and added a `Role` class that
has a smart property. Thus, we expect the DSL generation to eager load
the `Role` class and generate an RBI for it.

I made sure that this test breaks if we disable eager loading of
Zeitwerk dependencies.