Fix parameter names for attr writers with signature
Normally attr_writer generated methods do not have a name for the single
parameter that the method accepts. However, if the attr_writer has a
signature, then the parameter name should be the name of the attribute.

We have been defaulting the parameter name to `_` if the parameter name
didn't exist. But, as outlined above, if the writer method has a
signature, that ends up meaning that the parameter name in the signature
does not match the one in the method definition.

This commit fixes that by trying to detect writer methods with a
signature and naming the parameter properly in those cases.