Fix Sorbet ignore matching
Sorbet ignore matching does not exactly work like a regular expression
match. Instead, it tries to match whole path segments unless the ignore
path starts with a `/` character in which case it tries to match the
prefix of relative paths.

The new implementation first tries to get the relative path of each
found file to the working path. All matches are done against this
relative path. For each ignore string, if it starts with a `/`, the
first segment of the relative path. Otherwise, we expect it to match a
whole part of the relative path.