Copy IsolationHelper from ActiveSupport
Get rid of the parallel fork executor, so that we can provide a test
helper for user code that is trying to write a test for their own custom
DSL generator. This commit just adds the `IsolationHelper` module to
optionally provide the isolation to the test classes that want it. A
future change will add a test helper that consumers of Tapioca can
import into their test code as well.