Stop assuming that constants will have sensible hashes
Fixes #191

If a constant overrides the `hash` method at the constant level in an
incompatible way by returning a value that is not an `Integer`, then
`Set` stops working for that constant.

However, we were using a `Set` for filtering out ignorable ancestors
when discovering mixins, which was failing with a cryptic error message
if it ever came across a constant that did such a thing.

The solution is to use array difference rather than sets and
enumeration, which actually makes the code simpler in the process.