Change implementation to be O(NxM) :(
Damn you Ruby 2.4!

Apparently Ruby 2.4 uses `hash` internally to implement `Array#-` so I
had to fall back to checking the existence of each ancestor of the
constant in the ignorable ancestor array again, by comparing for equality
against each element of the ignorable ancestor list.

This causes the algorithm to be O(NxM) where N is the number of
ancestors of constant and M is the number of ignorable ancestors, but I
can't think of a cleaner way of doing this.