Streamline implementation again
I was not happy with the O(NxM) implementation from before and prompted
by @KaanOzkan and @Morriar's comments, I developed a faster implementation.

The new implementation is safer and more streamlined. I've added a new
helper method that returns inherited ancestors of a given class, which
we use in the main method to get the `object_id`s of the inherited
ancestors of the given constant and the given constant's singleton
class. We then check for the existence of the ancestors of the constant
(and its singleton class) against this set of ids to see which ones we
can omit.

This ensures that we are safe against any class overriding `hash`,
`eql?`, `equal?` or `==` on the class level. I've also added an extra
test to make sure that is the case.