Don't error on materializing missing specs
Some specs could be missing when trying to materialized the spec set,
for example, a gem could depend on another gem for one platform but not
depend on it for the current platform, which would make that dependency
exist in the lockfile but result in a missing spec for the current
platform.

The fix is to pass an empty array to `materialize` which would collect
the names of the gems that have missing specs.