Consolidate ActionControllerHelpers includes
At the moment when we're generating RBI with the ActionControllerHelpers DSL, we end up adding an include statement for every helper into every controller. This means that when someone adds a new helper module, the diff is really large because they're mutating every single controller's RBI file. While this is probably fine because they're generated files, it's still intimidating for folks to review.

This commit instead consolidates all of the included modules into one module, and then only includes that module into the controllers. That way when a new module is added, it only mutates the one RBI file. Admittedly this definitely has its shortcomings. Not every application is going to have an ApplicationHelper, so I don't know if we can necessarily rely on that. Ideally I'd like to generate one if one doesn't exist, but I couldn't find a way to do that with the current DSL infrastructure (gather_constants forces you into only generating RBI for existing files, even if a "helper" RBI module is necessary).