Purge unprocessed DSL RBI files
We collect the set of existing DSL RBI files in the filesystem and then
remove items from the set as files are generated. Whatever we are left
with in that set should be RBI files that are no longer needed.

The only complication for this strategy is the fact that the user can
supply a list of requested constant names, in which case, we need to
build the initial list as the filenames of the RBIs that those constant
names would have generated.