Change loader behaviour to support pre-Rails 6 apps
`Rails.autoloaders` did not exist in Rails pre-6.0, so we should not
assume that it will be there. We were faking this by defining an
`autoloaders` method on the fake `Rails` module defined in our test
repo, but we really should not be doing that.

Moreover, we were not actually trying to eager load a Rails app in a way
that is compatible with pre-6.0 versions, which this commit aims to add
as well.

Fixes: #239