Make generic type creation safe for sealed classes
Sealed classes refuse to be subclassed in any file outside the original
file they were declared in, but we use subclassing to represent concrete
generic types at runtime. So for a generic sealed type, we need to jump
through extra hoops to trick Sorbet runtime into thinking that what we
are doing is safe.