Extend protobuf support for map and repeated
Fields in protobuf can be marked as "repeated" or "map", which correspond to arrays and hashes, respectively. This commit extends our support for protobuf by correctly marking the types of these fields. It also adds an initialize method to each of the protobuf classes so that creating them can be typed.