Revert verbose command output upon verify failure
A previous PR had tried to make the command that was output when the
`--verify` step failed to be more surgical so that users can run a
command that would potentially be faster.

However, the constants that we collect as Tapioca is generating DSL RBI
files do not necessarily map to real constants inside the application.
Some of the constants are synthetic, for example
`GeneratedUrlHelpersModule`, and would not work if supplied on a
`tapioca dsl` command line. Thus, the verbose output ends up doing more
harm than good.