Treat default gems properly
If the version of a default gem specified in the `Gemfile` is the
version of that default gem that is bundled with the Ruby version, then
Bundler just uses the default gem from the Ruby library directory.

This prevents Tapioca from generating a proper RBI for the gem and we
end up generating an empty gem.

This commit makes sure that if a gem is marked as a default gem, we
check against the files listed in its gemspec inside the Ruby library
folder instead of looking inside what the gem install folder would have
been.