Fix app cache folder source gems
When Bundler is configured to cache gems inside the `vendor/cache`
folder in the app directory, for source gems, it creates git clones in
that folder. Those gems are consumed from those git clones, so they
look like they are part of the application code, as far as Tapioca is
concerned.

By adding an explicit check for the source files coming from
`Bundler.app_cache` folder, we can make sure we treat those source
locations are gem sources.