Add ability to query Sorbet features
We sometimes want to generate different code depending on which version
of Sorbet is in play. For that we need to check which version of Sorbet
we are working against.

In order to achieve this, we've added a query method for "features" for
Sorbet that does a minimum version check against the version of Sorbet
that Tapioca is running against. That allows features to turn themselves
on or off based on feature detection.

The first use of this query is in the ActiveSupport::Concern generator,
so that we don't end up generating `mixes_in_class_methods` calls with
multiple parameters on older versions of Sorbet.