`Loader` should not parse `Gemfile` unconditionally
The DSL run does not need to parse the `Gemfile` since it loads the
application eagerly, which is expected to be loading the `Gemfile`
anyway. However, `Loader` was assuming that it would always need to
materialize the gems listed in the `Gemfile`, which was wrong.

It is actually needed only for the `load_bundle` method, so it is better
for the bundle to be passed to that method only.