Strip leading `::` when adding to symbol queue
The symbol queue operates on the full names of constants and does not
know how to handle a name that starts with a `::`. If such a name is
added to the queue, it looks like it is an alias of a constant with the
same name but without the `::` prefix.