Use `to_ary_nil_support` Sorbet feature in AR relations
We should generate a `returns(NilClass)` signature for the model
`to_ary` method, since all models already inherit a sig-less `to_ary`
method from AR, which ends up making the return type of `to_ary` for
model classes `T.untyped`.

However, up to the version of Sorbet that supports `NilClass` return
type as special, we should not be generating that signature. Otherwise,
we would be making all flatten operations on model collections be typed
as `T::Array[NilClass]`.

Thus, we check for the version of Sorbet with that support and decide to
generate the `to_ary` method on the model with the proper signature.