Update `Tapioca::Error` class
As part of #536 and #238, we want to update the error messaging for DSL
compilation in `tapioca`. This introduces an updated `Tapioca::Error`
class which includes additional fields for `generator` (the Generator
which errored out) as well as `help_message` guiding messaging we can
provide tips in when we know what likely went wrong.

Signed-off-by: Cory Hutchison <cory.hutchison@shopify.com>