The `doc` option should not have a default value
The actual base default value of our flags is set in the `ConfigBuilder`
class. Any value passed to the command line flags (even if it is a
default value) will override all the values in the default and in the
config.

The upshot is that, since we had the default value of `false` for the
`doc` option, a `doc: true` addition to the config file had no effect
since the `false` coming from options would always override the config
one.