Add an extension of PrivateRelation for "where chain"
Rails 6.1 introduced an API for `Post.where.missing(:author)` that can
be used as syntactic sugar on top of joining and checking the associated
id is nil.

This leverages the existing QueryMethods::WhereChain class which we
haven't previously used in Tapioca.

WhereChain should be returned from `#where` and should give the ability
to do where.missing, where.not (and soon, where.associated). The result
should give back a plain PrivateRelation because
where.missing().missing() is invalid.