Handle eval'ed classes in TracePoint properly
For eval'ed classes/modules, the `path` attribute of the TracePoint
object ends up being `"(eval)"`, which does not help us figure out
where the class/module is being defined.

Instead, we can look back up the call stack until we find the first
caller location that has a `path` that is not equal to `"(eval)"` which
is where the call is ultimately being made from.