Fix doc generation on CI
If the `config` DSL generator was loaded after `identity_cache` DSL
generator, then we had a failure because `Rails::Railtie` would be
loaded but there wouldn't be a properly configured application around.

The problem was because `identity_cache` was requiring `rails/railtie`
explicitly, which it shouldn't do. Any proper installation of IDC should
be inside of a properly configured Rails application or, alternatively,
it would be loaded inside a non-Rails application which would not
trigger the IDC Railtie anyway.

The best fix is to move the require of `rails/railtie` to IDC generator
tests instead.