Exclude anonymous or value constants from DSL generation
Fix #592

The whole idea of DSL generation revolves around being able to generate
files that are named after the constants that are being decorated, so
it makes sense to filter out any constants that were collected that are
anonymous.

Moreover, while we enforce that the constants that we expect to be
collected are supposed to be classes/modules, we never actually enforce
that at runtime, in the absence of Sorbet runtime, so I've added a
filter to exclude value constants from the list of collected constants
as well.