Add pessimistic version dependencies on our gems
Given that we are likely to break API compatibility often in RBI or
Spoom, we need a way for Tapioca to not allow people to install
incompatible versions of those gems. For that reason, we should add
pessimistic version dependencies on those gems (as described here:
https://guides.rubygems.org/patterns/#pessimistic-version-constraint),
so that RBI and Spoom can bump the minor version when they make an
incompatible release and Tapioca will not automatically allow end users
to install those versions.