Refactor config infra to be integrated with Thor
Instead of relying on a strongly typed struct class like `Config` to
read the supplied options, we decided to instead work with Thor and not
against it. We've accepted that getting strong types for options will be
hard, which has allowed us to hook into the `options` call on Thor, in
order to perform the `default` -> `config` -> `command line` supplied
options merge in that order.

The major problem to solve was how Thor works with default values
supplied to options. Thor sets an option value to the default value if
the user has not supplied the option on the command line. This prevents
the application from being able to tell apart the cases when the option
was supplied by the user but it was supplied as the default value, or
if the option was not supplied at all, and it was set to the default
value.

We need to be able to differentiate between the two cases, so that we
can override default, non-supplied values by the values supplied in a
config file.

To be able to pull this off, we intercept the `initialize` method of
the `Thor` subclass and filter out the default values supplied to
options. We store those default values in a separate hash, so that when
the application asks for the options for the command, we can do the
merging of the 3 levels of values in the correct way and resolve the
option to the right value.

Co-authored-by: Alexandre Terrasa <583144+Morriar@users.noreply.github.com>