Improve default gem test
The following improvements have been made:

1. The default gem name that is tested against is made dynamic so that
we can change which gem we test against more easily in the future. The
`rdoc` gem we were testing with wasn't the best candidate, since we
already use it in a different test, which might make a Rubygems version
of it (with the same version) to be installed on the system before this
test runs. That would make that version be selected instead of the
default version on the system (even if they have the same version).
2. We hint to Bundler with the exact version of the default gem that is
installed on the system, so that Bundler does not try to install a newer
version of the gem from Rubygems. This version discovery also makes it
easy to figure out the generated RBI filename as well, instead of
resorting to Regexp matches.