Fix typed override `Hash` defaults
The new merge options was not doing the correct thing when calling
`merge!` on `result`, since the result is now created as an indifferent
access hash, which doesn't seem to implement the block callback for
`merge!`.

This was causing the default `"activesupport" => "false"` setting to be
lost.

Instead we create a normal Hash and then convert it to an indifferent
access hash instead.