Refactor how `SmartProperties` methods are generated
We were generating `SmartProperties` methods on the constant definition
itself, but this is not how things are implemented in Smart Properties
and it makes it hard to decide which method to generate.

By generating Smart Properties defined methods in an inner module, we
model the actual operation of what really happens at runtime better.
Moreover, we don't need to keep doing checks to see if the method was
overridden in the constant itself, to decide if we should generate the
method or not.