Turn `TypeVariable` into a `Module` so that it auto binds to its name
One of the problems we've had was how (and when) to figure out that a
type variable (for example, constructed by `type_member`) is bound to
a specific constant name (for example, to `Elem`). We were resorting to
scanning the subconstants of a constant as we were processing it, but
that made the process dependent on when we processed a certain constant
and would break when there was a type variable access to a constant that
was not processed yet.

However, we know that modules are automatically bound to the name of the
constant they are assigned to at the time of assignment. For example:

```ruby
foo = Module.new
foo.name # => nil
Elem = foo
foo.name # => "Elem"
```

So, there realization is that if we make `Tapioca::TypeVariable`
instances modules, then we wouldn't have to worry about this since the
name binding would be done by Ruby automatically at assignment time.

All we have to do is to make `Tapioca::TypeVariable` subclass from
`Module` so that its instances are modules, and to special case that
module when we are processing constants, which is exactly what this
commit does.