Start using compared by identity Set operations
When doing a difference of arrays for finding interesting ancestors,
we need to be careful in case a constant has overridden `Object#hash`.
Since array subtractions use `Object#hash` to determine equality, this
operation will fail.

Instead, we need to use a Set with compare-by-identity semantics, so
that `Object#hash` is never called.