Set type of CollectionProxy to T.untyped
CollectionProxy doesn't allow bare uses, which is a problem for gems
like ActionText. Using a bare CollectionProxy results in:

> Malformed type declaration. Generic class without type arguments

This means that as soon as you enable ActionText, whether you're using
it or not, you're unable to have `srb tc` pass.

By explicitly stating `T.untyped`, we no longer see this issue.