Add descriptive message for unexpected DSL exceptions
Currently, when a DSL compiler raises an error while running a `decorate`
operation, it is not very obvious which compiler failed and during the
processing of which constant. The failing DSL compiler is kind of obvious
if one looks at the backtrace, but the constant that it is failing on
is very hard to understand if one is not running DSL generation in verbose
mode.

This PR handles unexpected exceptions, shows a friendly error message for
them and re-raises the exception to keep the current behaviour.