Clean traces of Active Support before loading gem pipeline tests
It seems Active Support 7.0+ loads a lot of the library when using the
test runner, which makes a lot of AS classes appear as loaded in Tapioca
which is a problem for the test expectations.

The "fix" is to remove ::ActiveSupport as a constant and remove all
traces of "active_support" from loaded features, so that a call to
`require "active_support"` will go through properly.