Fix type of `column_type` parameter for `handle_unknown_type` method
The passed value of `column_type` parameter is not necessarily always
an `Object`. For example, `ActiveSupport::Types::Serialized` does not
have `Object` as an ancestor, since it is defined as:
```
class Serialized < DelegateClass(ActiveModel::Type::Value)
```
which makes it subclass from `BasicObject`.

We were using `Object` as the type for `column_type` parameter, but we
really wanted to use `BasicObject` instead.

I added a serialized column to a test case to see this fail before the
change and pass after the change.