Load Rails engine symbols properly
Since [this commit](https://github.com/Shopify/tapioca/commit/ed9e590936cbd7cd5283c7bbb20f1b87316d23b4)
we have not been loading Rails engine symbols from gems. The commit
started filtering Rails engines by looking up the name of the engine
in the payload symbol table which was not the correct thing to do and
would never return an engine for consideration.

What should be happening is to find the engine that is exported from a
gem, and to collect all the files that belong to that engine, so that
Sorbet can find the symbols.

Generally, this would not be causing a problem if all the constants
loaded from an engine were under the gem's top level namespace, since we
walk the namespace down and would discover those constants (since there
was, hopefully, no problem with loading the engine files). If the engine
was exporting a top-level constant, though, we would not be able to catch
that since we never scanned the symbols. This was happening, for example,
for the `view_component` gem's [`PreviewHelper` module](https://github.com/github/view_component/blob/8dd9063d0b74125c0e20222ee76e64faac6d13a2/app/helpers/preview_helper.rb).