Remove `T.unsafe` from extension method call
Sorbet doesn't know that `::FrozenRecord::Base` will have
`::Tapioca::Dsl::Compilers::Extensions::FrozenRecord` prepended to its
`singleton_class`. Therefore, it does not allow us to call the methods provided
by that module.

If we tell it that `ConstantType` will be the union of both, it allows us to do so.