New approach to finding attached class
Historically, to identify the attached class of a singleton class, we've
parsed the name (result of the #inspect method) of the singleton class
and then constantized the result.

However, this doesn't work when the #inspect method of a singleton class
is overridden because the return value of #inspect no longer conforms to
the structure we expected.

Instead, we can search ObjectSpace for the class that has the correct
singleton class. This is less performant but more correct.