Change how we get file candidates from `Pry::WrappedModule`
It turns out that Pry does a lot of extra work to get a list of
candidates for a given wrapped module. One of these extra things is to
attribute the methods that are introspected to the constant that defines
them in the source file. This is not only a performance issue, but it
was also failing for methods added to `Object` by being defined at the
top level, and was resolving their `file`s as `nil`.

Since we don't need the machinery that the `candidates` method gives us,
we can just rely on the `method_candidates` method (alas, it is a
private method), to grab the unique file names that have defined methods
on the wrapped module.

The `method_candidates` returns an array of arrays of `Pry::Method`
objects. We flatten it into a single list, grab unique source file names
from it. This should give us the proper list of file candidates.

This also reduces our dependency on `Pry:WrappedModule` and allows us to
reimplement it inside Tapioca in the near future.