Filter ActiveSupport `String` extensions in tests
A recent [change in Rails][0] has stopped loading core extensions to
`String` very early. These tests were failing because of that change
on Rails main builds.

The best solution to this is to filter those constants in the compiled
output, since these constants are not primary things being tested in
those tests anyway.

[0]: https://github.com/rails/rails/commit/0170745b376acd150fec5f8cc57253cc1ffe0cf2