Start using alias-method-chain in mixin tracker
The current prepended-module way of intercepting mixin events does not
work if another gem uses alias-method-chain to intercept the same events.

For example, the [`blankslate` gem uses alias-method-chain to intercept
`append_features`][0] and if that code is loaded after the mixin tracker
then it causes an infinite loop.

The reason why the infinite loop happens is the following sequence of
events:

1. Mixin tracker prepends a module onto `Module` with the `append_features`
  method.
2. `blankslate` is loaded and uses alias-method-chain to intercept
  `append_features` on `Module`. In order to do that, it stores a
  reference to the original `append_features` method, but at this
  point, that "original method" is the one from the mixin tracker.
3. When `append_features` method is called, it called the method on the
  prepended module first, so the mixin tracker's `append_features` method
  gets called.
4. The mixin tracker's `append_features` method calls `super` which is
  the one from `blankslate`.
5. `blankslate`'s `append_features` method calls the stored "original
  method" which is the mixin tracker's `append_features` method.
6. Go to step 3.

In this case, it is better to use alias-method-chain in mixin tracker,
as well.

[0]: https://github.com/masover/blankslate/blob/master/lib/blankslate.rb#L101-L102