Add support for non-generic WeakMap in pipeline
Adds support for using `ObjectSpace::WeakMap` without
a generic type argument. This change is guarded by a
check for the availability of a specific Sorbet
feature (`non_generic_weak_map`), which is required
for this change to be applied. This update allows
Tapioca to use the non-generic version of `WeakMap`
when running on a version of Sorbet that supports it,
while still providing support for earlier versions that
require the generic type argument.