Ensure that trackers stay enabled until after `gem` command is run
Since all commands other than `gem` and `dsl` commands now inherit from
`CommandWithoutTracker`, as soon as we invoke the `configure` command
inside `init`, we ended up disabling the trackers. This meant that by
the time the `gem` command got to run, the trackers were already
disabled and that we generated the RBIs with missing information.

The fix is to create a block in which trackers can't be disabled and
using that up to and past the point where the `gem` command is run.