Simplify check for GraphQL defined method
Any method defined by `GraphQL::Schema::InputObject` will end up
having a source file that points to the same file that defines the
`GraphQL::Schema::InputObject.argument` method. So, we can simplify
the check for whether a method is defined by
`GraphQL::Schema::InputObject` or if that method was redefined later,
by comparing the source file of the method to the source file of the
`GraphQL::Schema::InputObject.argument` method.