Stop trying to mimic Rails and do the simpler Zeitwerk autoloading
It is simpler to iterate over all the engines and add their eager load
directories to a fresh Zeitwerk::Loader instance.

When doing this, we need to be careful to not add a directory to our
loader that is already managed by another loader. So, we collect all the
directories that are already managed by all existing loaders and
skip those directories when adding them to our loader.