Don't ignore default gems if Ruby vendored in app
This is similar in spirit to https://github.com/Shopify/tapioca/pull/34,
except that it's probably a much less common scenario.

At GitHub we vendor a Ruby build within our app directory, which means
that our default gems meet the conditions of the `gem_in_app_dir?`
method—they are within the app, but not in the bundle directory. That
results in these default gems getting ignored.

We also sometimes generate PRs via GitHub actions, using Ruby from the
setup-ruby action. In these scenarios the default gems are NOT within
the app, and are thus NOT ignored. To get around having PRs that
continuously add and remove these rbis, we ended up explicitly excluding
them.

This commit aims to make the behavior consistent regardless of where
Ruby is installed. It does that by checking whether the gem path is
inside of Ruby.

Co-authored-by: John Hawthorn <jhawthorn@github.com>