If a message class does not have a name, map it to `T.untyped`
If the `msgclass` of a subtype is not assigned to a constant, then its `#name` method would return `nil` and the compiler would return it as the type name of the descriptor. Not only does this break the contract of the `type_of` method defined by its signature which promises to always return a `String` instance, but it also leads to us generating broken RBI files, since these `nil` values serialize to empty strings.

 This commit fixes this by mapping the message class to `T.untyped` if it does not have a name.