Replace direct `include Kernel` by `requires_ancestor { Kernel }`.
Since #1530, inclusion of Kernel is properly generated by Tapioca.
When generating RBIs, this creates an issue for Isolation modules
where the `requires_ancestor { Kernel }` on `DslCompiler` is
reported as useless.

we do not have this error locally in Tapioca because the include
is conditional so Sorbet doesn't see it. But we get it with RBIs
in other projects because Tapioca finds the include through
reflection.

The Isolation sub-modules do not need to include Kernel.
It seems required only to call fork and exit so we can replace
them by `requires_ancestor { Kernel }`.

Closes https://github.com/Shopify/tapioca/issues/1563.

Signed-off-by: Alexandre Terrasa <alexandre.terrasa@shopify.com>