Extract multiple commands from `Dsl` command
We'd been using the `Dsl` command to perform three distinct tasks:

1. generate DSL RBI files for specified or all constants/paths
2. verify that the DSL RBI files are up-to-date
3. list DSL compilers

All of this functionality bunched together in a single command was making it hard to change behaviour or to add uniform behaviour to all `Command` classes, since not all of them would call the execute method in the end.

This commit extracts an abstract dsl command super class and splits the previous `Dsl` command into separate commands that each handle one of these three tasks.