Dealias constant prefixes before writing to RBI
Tapioca is currently generating RBIs with type checking errors when a
constant is aliased and then that alias is used as part of the namespace
of another constant. e.g.

```
module Foo; end
module Bar; end

F = Foo
F::B = Bar
```

If we do not dealias the constant `Foo` before writing `F::B = Bar`,
Sorbet will think we are trying to redefine the alias `F` and will raise
a type checking error.

This commit updates the pipeline so that `F::B = Bar` will be
transformed to `Foo::B = Bar` before writing to the RBI, which will
not create any new type checking errors.

Co-authored-by: Alexandre Terrasa <alexandre.terrasa@shopify.com>