Deprecate the `todo` command
The `todo` command had a valid use case when Tapioca was newer and
couldn't generate RBI for constants in some cases. However, now that
Tapioca is more mature, `todo` is more likely to cover up a type
checking issue than fix it.

This commit adds a deprecation message to the `todo` command warning
users that the command will be removed entirely in a future version.