Reproduce issue with URI method RBI in ruby-lsp gem
v0.8.0 of the ruby-lsp gem added a URI constant with some method
overrides. When tapioca generates RBI for this version of the gem, it
does not include RBI for those methods defined on the URI constant.

This commit adds two tests: one that demonstrates the failure on the
ruby-lsp gem, and the other that attempts to reproduce the failure on a
mock gem -- but it's passing 🤷