Command suggestion includes environment args
Different Rails environments can yield different RBIs when running
bin/tapioca dsl. An environment argument was added to `dsl` to overcome
this problem.

This commit changes the suggestion in the error message from
`bin/tapioca dsl --verify` so that it explicitly suggests using the
environment argument that caused the verification to fail.

This is difference is specially important for troubleshooting in CI
environments, where `bin/tapioca dsl` without `-e test` might not fix
the issue.