Enhance type inference for ActiveModel attributes
This commit improves the `type_for` method in the
`ActiveModelAttributes` compiler. Previously, if an attribute's type was
 not one of the predefined types (Boolean, Date, DateTime, Time,
 Decimal, Float, Integer, String), it was automatically assigned as
 `T.untyped`.

With this change, the method now checks if the attribute type is an
instance of `ActiveModel::Type::Value`. If it is, it returns the class
name of the attribute type. This allows for better type inference for
custom attribute types.

If the attribute type is not an instance of `ActiveModel::Type::Value`
and is not one of the predefined types, it still defaults to `T.untyped`

This enhancement provides more accurate type information in the
generated RBI files, especially for applications that use custom
attribute types.