Update the workaround for `debug` gem
The current approach of disabling `debug` loading makes it impossible to
debug Tapioca itself with the gem. But by lazily activate the debugger, we can
avoid the side-effect as well as allowing breakpoints to lazily load the
debugger.

It's the same as doing `require "debug/prelude"`, which only defines the
breakpoint methods.