Stop requiring stuff from DSL compilers
DSL compilers should not change the object space of the application they run in, and thus should not really be doing any requires. This commit removes all requires from DSL compilers, and instead relies on the application to require the necessary files.

A DSL compiler should be responsible for checking if the constants it is expecting to have been loaded are actually loaded, and if not, it should not be defining the DSL compiler class at all.