Update ActiveSupportCurrentAttributes compiler to include block param on getters
For gems that use `ActiveSupport::CurrentAttributes`, the tapioca gem
compiler now includes a block param in the attribute getters. However,
the DSL compiler does not include this block param, which causes type
checking errors.

This modifies the DSL compiler to add block params to
the attribute getters on `ActiveSupport::CurrentAttributes` classes.