Stronger checks for existence of `config` gem
Just checking for the existence of a `Config` constant does not necessarily guarantee that the `config` gem is loaded, since an application could define their own top-level `Config` constant.

It is better to check for the existence of `Config::VERSION` and `Config.const_name` as well, to be reasonably sure that the constant is coming from the gem.