Change signature of find to accept specific types instead of T.untyped
The problem is that T.untyped is both the superclass and the subclass of all types in Sorbet, so that when given an untyped parameter the first signature (returning an array) would be selected whereas probably a single model instance would be intended (see the removed note in a previous commit about find being untyped in the first place).

Integer and String (think UUID) should cover the most common cases, but it is technically possible to pass anything that can be quoted by Active Record except nil (which is disallowed by ActiveRecord::FinderMethods#find_with_ids).