Add ripper DSL
This doesn't quite fit in to the normal idea of a DSL, but I think
it fits best here. Basically, because ripper isn't a normal gem and
doesn't have a version, there's not a great way to generate static
RBIs for it, because it depends on your current version of Ruby.