Refactor DSL compiler options to be a flat namespaced hash
I've been thinking about this for a while, and I think namespacing compiler options was a little heavy handed. I've been thinking about cases where a single option might need to be consumed by multiple Active Record compilers, so I think it makes sense to have a flat hash of options that can be consumed by any compiler.

This makes the code simpler and makes it easier to pass these options on the command line interface, as well.