Generate correct return types in IdentityCache compiler
There is a difference in behavior between the return types of the existing
IdentityCache compiler and how IdentityCache compiler actually works.

In the aliased fetch methods, IdentityCache will return an array unless
the field has the `unique` property, even if it is not a multi-type.
Previously, our IdentityCache compiler was always setting the return type
as a single object for those methods.