Make DSL compiler test helper work without needing an extra require
Currently Tapioca exports a DSL compiler test helper that is used by other app or gems to test their DSL compilers. However, the helper requires a lot of Tapioca DSL classes to be defined before it can be used. This is a problem because it means that the helper can't be used in isolation, without the rest of Tapioca. A common workaround for this is to start requiring "tapioca/internal" in the test file that uses the helper, but that ends up causing Sorbet runtime related behaviour changes due to the Sorbet runtime patches that that require pulls in.

In order to fix this, all we need to do is to require "tapioca/dsl" which brings in only the DSL classes that the helper needs. This way, the helper can be used in isolation without needing to require "tapioca/internal" and without causing any Sorbet runtime related behaviour changes.