Make it possible to extend list of id types for ActiveRecord relations
By moving the list of id types to a constant, we are now making it possible to extend the list of id types for ActiveRecord relations in gem or application level "compilers".

All that a "compiler" needs to do is to read the list of id types from the `Tapioca::DSL::Compilers::ActiveRecordRelations::ID_TYPES` constant, add the new id type to the list and then assign the result back to `Tapioca::DSL::Compilers::ActiveRecordRelations::ID_TYPES`.