Fix missing gem spec test
Our test for skipping RBI for missing gem specs was broken by a change
in how bundler represents specs for default gems
(https://github.com/rubygems/rubygems/pull/7673), and thus changes
when they're considered "missing." Our test happened to use a default
gem, which is why it broke. However, the test doesn't rely on
the example being a default gem, so we can change it to a different
gem, which will fix the test.