Add Tapioca Addon gem RBI generation support
To support gem RBI generation, we needed a way to detect changes in
Gemfile.lock. Currently, changes to this file cause the Ruby LSP to
restart, resulting in loss of access to any previous state information.

By running git diff on Gemfile.lock, we can detect changes to the file,
and trigger the gem RBI generation process.

Then we can parse the diff output to determine which gems have been
removed, added or modified, and either remove the corresponding RBI
files or trigger the RBI generation process for the gems.