Add Tapioca Addon gem RBI generation support
To support gem RBI generation, we needed a way to detect changes in
Gemfile.lock. Currently, changes to this file cause the Ruby LSP to
restart, resulting in loss of access to any previous state information.

By running git diff on Gemfile.lock, we can detect changes to the file,
and trigger the gem RBI generation process.

When changes are detected, we parse the output of git diff to identify
added, modified, or removed gems. We then execute `tapioca gem` command
for added, or modified gems, and remove rbi files for removed gems.