Use method overloading to improve WhereChain definition
`ActiveRecord::QueryMethods#where` returns a `WhereChain` instance only when it is called with no arguments. When called with arguments, it returns the same `ActiveRecord::Relation` instance.

Previously, we didn't have proper overloading support in RBI files, so the best thing we could do was act like `WhereChain` was a subclass of the related relation class. This was not ideal because it would allow users to call methods on `WhereChain` that were not actually defined on it. For example, one could do `User.all.where(name: "John").not` which is not valid.

Now that we have support for overloading, we can define `where` to return a `WhereChain` instance only when called with no arguments, and in all other cases, we can say that it returns the same relation instance.