[DO NOT MERGE] cherry-pick Bundler 2.6 support
Fix specs to play nice with Bundler 2.5.23

Gem::Specification.stubs used to be a `Bundler::SpecSet` when
`bundler/setup` has been required. However, that was unintentional and
was reverted to be a plain array of specs like with standard RubyGems in
Bundler 2.5.23.

This change make specs support that change.

Remove unnecessary Bundler input

RubyGems 3.5.x ships with Bundler 2.5.x, so it's unnecessary.

Sync locked Bundler and Bundler test in CI to 2.5.23

In CI, we bump to RubyGems 3.5.23 which ships with Bundler 2.5.23.

Support Bundler 2.6

Bundler 2.6 made a change where missing_specs are now kept in the
original resolve spec set, not in the spec set with materialized
specifications.

This change adapts tapioca to that.