Stop ignoring symbols in payload when discovering subconstants
This change makes Tapioca consider subconstants of symbols that may already be defined in the payload, ensuring that all relevant subconstants are discovered and processed correctly.

This prevents a case where a new subconstant is defined in a default gem that doesn't get picked up by the payload yet, causing Tapioca to miss it and not generate the corresponding RBI file.