Silence warnings when eager loading autoloaded constants
Some gems print a warning when particular files are loaded, since some dependencies that they might have been expecting to be loaded are not. These warnings are not useful when Tapioca is trying to eager load these constants, and they break our tests.

In particular, the `prism` gem checks for the availability of the `parser` gem and `ruby_parser` gem, and prints a warning if they are not available.

This commit silences warnings around the eager load operation.