Fix handling of anonymous types in generic arguments
If the argument to a generic type was an anonymous type, the name of the type as returned by our `T::Types::Simple#name` patch was `nil`. But that meant that Tapioca was generating a type that was broken, like `T::Array[]`.

The right thing to do is to return `T.untyped` for anonymous types, and not `nil`. So that we at least have a valid type, like `T::Array[::T.untyped]`, if not the original type.