Look at sections from the back in case there are warnings at the front
Upgrading sorbet (which happens in CI when the Gemfile.lock is
removed) currently results in a Warning.

Rather than trying to work around whether the warning is missing or
present in a section by itself, or present in a section we are looking at,
just count from the back instead of the front.